requirements.txt
dhanhq

openpyxl

pandas
psycopg2-binary
python-dotenv

websockets